requirements.txt
scikit_learn
termcolor
torch~=2.2.0
torchvision~=0.17.0+cu121
torchaudio==0.11.0
torchtext==0.12.0
tqdm~=4.66.4
transformers
numpy~=1.26.3
scikit-learn~=1.5.0
pillow~=10.2.0
pandas~=2.2.2
einops~=0.8.0
timm~=1.0.3
safetensors~=0.4.3
setuptools~=65.5.0
lora~=0.1
torchsummary~=1.5.1
datasets~=2.20.0
PyYAML~=6.0.1
wandb~=0.17.1